Find neovim lua api functions in help
Add more lookups so when we're looking up neovim's vim wrappers, we find
the correct help.

Using ' prefix for options since that's included in their tags: for
vim.o.ea we want to find 'equalalways' and not :ea.

Add whole word lookup to ensure we find functions over commands. Vim
help seems to generally prioritize commands, so where we know it's a
function we want to find that help by including that trailing (.

Test
    vim.bo.iskeyword
    vim.o.ea
    vim.g.fugitive_dynamic_colors
    vim.api.nvim_buf_set_lines()
    vim.fn.substitute()